requirements.txt
# Core Dependencies
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2
email-validator==2.1.0.post1
fastapi==0.109.2
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
pytest==7.4.3
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Core ML Stack - Pinned Versions
torch==2.1.2
transformers==4.36.2
huggingface-hub==0.19.4  # Specific version known to work with transformers 4.36.2
safetensors==0.4.3

# Runtime Dependencies
numpy==1.26.0
python-json-logger==2.0.7
jinja2==3.1.2
aiofiles==23.2.1

# AI/ML Specific
airllm==2.3.0  # Pinned to specific version
accelerate==0.25.0  # Pinned to specific version
openai==1.3.7  # For OpenAI-compatible API

# CLI Dependencies
typer[all]==0.9.0
rich==13.7.0
docker==7.0.0  # For Docker SDK
python-dotenv==1.0.0  # For loading .env files
requests==2.31.0  # For HTTP requests

# Additional dependencies
protobuf==3.20.3
tokenizers==0.15.0  # Specific version compatible with transformers 4.36.2
packaging==23.1

# Explicit dependencies to avoid conflicts
filelock==3.12.4

# Additional ML dependencies
regex==2023.12.25
sentencepiece==0.1.99
